Refactor stencil rendering to allow for clearing.
