+++ include/mbgl/mtl/mtl_fwd.hpp
@@ -22,6 +22,7 @@ class Buffer;
 class CommandBuffer;
 class CommandQueue;
 class Device;
+class DepthStencilState;
 class Function;
 class RenderCommandEncoder;
 class RenderPipelineState;
@@ -45,6 +46,7 @@ using MTLBufferPtr = NS::SharedPtr<MTL::Buffer>;
 using MTLCommandBufferPtr = NS::SharedPtr<MTL::CommandBuffer>;
 using MTLCommandQueuePtr = NS::SharedPtr<MTL::CommandQueue>;
 using MTLDevicePtr = NS::SharedPtr<MTL::Device>;
+using MTLDepthStencilStatePtr = NS::SharedPtr<MTL::DepthStencilState>;
 using MTLFunctionPtr = NS::SharedPtr<MTL::Function>;
 using MTLRenderCommandEncoderPtr = NS::SharedPtr<MTL::RenderCommandEncoder>;
 using MTLRenderPassDescriptorPtr = NS::SharedPtr<MTL::RenderPassDescriptor>;
