+++ src/mbgl/renderer/paint_parameters.hpp
@@ -118,7 +118,7 @@ private:
     template <typename TIter>
     using GetTileIDFunc = std::function<const UnwrappedTileID&(const typename TIter::value_type&)>;
     template <typename TIter>
-    void renderTileClippingMasks(TIter beg, TIter end, GetTileIDFunc<TIter>&&);
+    void renderTileClippingMasks(TIter beg, TIter end, GetTileIDFunc<TIter>&&, bool clear);
 
     // This needs to be an ordered map so that we have the same order as the renderTiles.
     std::map<UnwrappedTileID, int32_t> tileClippingMaskIDs;
